mods

+++ Solver/Model/Day9Input.cs
@@ -6,14 +6,14 @@ namespace Solver.Model
 {
 	public class Day9Input : IInput<Day9Input>
 	{
-		public int[] Commands { get; set; }
-		public int Input { get; set; }
+		public long[] Commands { get; set; }
+		public long Input { get; set; }
 
 		public Day9Input Parse(string[] values)
 		{
 			Commands = values.First()
 							 .Split(',')
-							 .Select(int.Parse)
+							 .Select(long.Parse)
 							 .ToArray();
 
 			return this;
